Splitted tests.py into multiple test files in baskets
